Move database CRUD functions to shared module.
