+++ shared/database/database.go
@@ -1,7 +1,9 @@
-package main
+// Package database provides CRUD operations for the MySQL Buddhabrot database.
+package database
 
 import (
 	"database/sql"
+	"fmt"
 	"os"
 
 	"github.com/go-sql-driver/mysql"
@@ -31,7 +33,9 @@ func connect() (*sql.DB, error) {
 	return db, nil
 }
 
-func insert(json string) (int64, error) {
+// Insert inserts a plot record with the given JSON Buddhabrot parameters, and
+// returns the ID of the newly created plot.
+func Insert(json string) (int64, error) {
 	var err error
 
 	var db *sql.DB
@@ -51,7 +55,8 @@ func insert(json string) (int64, error) {
 	return id, nil
 }
 
-func getFilename(id int64) (string, error) {
+// GetFilename retrieves the filename associated with the given plot ID.
+func GetFilename(id int64) (string, error) {
 	var db *sql.DB
 	db, err := connect()
 	if err != nil {
@@ -65,4 +70,27 @@ func getFilename(id int64) (string, error) {
 	}
 
 	return filename, nil
+}
+
+// Update updates the filename associated with the given plot ID.
+func Update(id int64, filename string) error {
+	var err error
+
+	var db *sql.DB
+	db, err = connect()
+	if err != nil {
+		return err
+	}
+
+	result, err := db.Exec("UPDATE plots SET pngfile = ? WHERE id = ?",
+		filename, id)
+	if err != nil {
+		return err
+	}
+	if rows, err := result.RowsAffected(); err != nil {
+		return err
+	} else if rows != 1 {
+		return fmt.Errorf("update affected %d rows", rows)
+	}
+	return nil
 }
